Fixed output
Now it needs less time to make a table in researches

+++ 1/head.h
@@ -12,4 +12,4 @@ using std::endl;
 
 
 typedef float real;
-typedef float real_sum;
+typedef double real_sum;
